requirements.txt
requests
tqdm
scikit-learn
pandas
seaborn
matplotlib
hydra-core
black
pylint
commitizen
